Cleanup webhook logic
